Moving connect tests back to bindings tests
